requirements.txt
Django
Markdown
requests
gunicorn
pyodbc
python-dotenv
django_redis
tabulate